+++ emcc.py
@@ -842,10 +842,10 @@ def get_cflags(options, user_args):
     cflags.append('-fvisibility=default')
 
   if settings.LTO:
-    cflags.append('-flto=' + settings.LTO)
+    if not any(a.startswith('-flto') for a in user_args):
+      cflags.append('-flto=' + settings.LTO)
   else:
-    # With LTO mode these args get passed instead
-    # at link time when the backend runs.
+    # In LTO mode these args get passed instead at link time when the backend runs.
     for a in building.llvm_backend_args():
       cflags += ['-mllvm', a]
 
@@ -1842,7 +1842,6 @@ def phase_linker_setup(options, state, newargs, settings_map):
       '___emscripten_pthread_data_constructor',
       '___pthread_tsd_run_dtors',
       '__emscripten_call_on_thread',
-      '__emscripten_do_dispatch_to_thread',
       '__emscripten_main_thread_futex',
       '__emscripten_thread_init',
       '_emscripten_current_thread_process_queued_calls',
@@ -1858,6 +1857,7 @@ def phase_linker_setup(options, state, newargs, settings_map):
       '_emscripten_sync_run_in_main_thread_4',
       '_emscripten_tls_init',
       '_pthread_self',
+      '_pthread_testcancel',
     ]
     # Some of these symbols are using by worker.js but otherwise unreferenced.
     # Because emitDCEGraph only considered the main js file, and not worker.js